requirements.txt
yt-dlp
ttkbootstrap
Pillow
requests